Re-order the examples
Until now, we were mentioning all the bad ones together
and all the good ones together!

However, it appears to be much more neater and readable
if we alternate b/w the bad one and the good one.
This way, we can clearly express the bad part in the bad
example and how could it be tweaked to make it "good"!

Signed-off-by: Utkarsh Gupta <utkarsh@debian.org>